==> 2.HousePrediction-GNAF.ipynb <==
# Price Estimation Model
### Alireza Shirkhodaee, Arian Fallah, Kamyab Rajabizadeh

# import libraries

import pandas as pd
import numpy as np
import xgboost as xgb
import pickle
import warnings
warnings.filterwarnings("ignore")

# estimation model

def price_estimation(X):
    with open("xgboost_model.pkl", "rb") as f:
        xgb_reg_loaded = pickle.load(f)
    Y = xgb_reg_loaded.predict(X)
    df_pred = pd.DataFrame({'y_pred': Y})
    return df_pred

# GNAF

import pandas as pd

excel_file_path = r"d:\Users\A.Fallah\Desktop\Price Estimation Model\From Shirkhodaei\Gnaf_final.csv"
df_gnaf = pd.read_csv(excel_file_path)
df_gnaf

df_gnaf = df_gnaf.drop("unit_per_floor", axis=1)
df_gnaf = df_gnaf.rename(columns={'Neighborhood': 'neighborhood'})
df_gnaf = df_gnaf.rename(columns={'unit_area': 'area'})
df_gnaf = df_gnaf.rename(columns={'floor': 'floors_count'})
df_gnaf = df_gnaf.rename(columns={'mode_unit_per_floor': 'unit_per_floor'})
df_gnaf = df_gnaf.drop(columns=['unit'])

df_gnaf = df_gnaf.dropna(subset=['neighborhood', "area", "floors_count", "unit_per_floor"])

df_gnaf

excel_file_path = r"d:\Users\A.Fallah\Desktop\Price Estimation Model\District.xlsx"
df_dis = pd.read_excel(excel_file_path)
df_dis = df_dis.rename(columns={'Postal Code': 'Tours'})
df_dis['Tours'] = df_dis['Tours'].astype(str)
df_dis

df_gnaf['postal_code'] = df_gnaf['postal_code'].astype(str)
df_gnaf['Tours'] = df_gnaf['postal_code'].apply(lambda x: x[:5])
df_gnaf = pd.merge(df_gnaf, df_dis, on='Tours', how='left')
df_gnaf

df_avg = pd.read_csv(r"d:\Users\A.Fallah\Desktop\Price Estimation Model\Average-Neighborhood.csv")
df_avg = df_avg[["neighborhood", "district", "weighted_average_price"]]
df_avg = df_avg.rename(columns={'district': 'District'})
df_avg

df_gnaf = pd.merge(df_gnaf, df_avg, on=['neighborhood','District'], how='left')
df_gnaf['weighted_average_price'] = df_gnaf['weighted_average_price'].astype(np.int64)
df_gnaf['floors_count'] = df_gnaf['floors_count'].astype(np.int64)
df_gnaf['unit_per_floor'] = df_gnaf['unit_per_floor'].astype(np.int64)
df_gnaf['area'] = df_gnaf['area'].round().astype('int64')
df_gnaf["postal_code"] = df_gnaf["postal_code"].astype(np.int64)  

df_gnaf = df_gnaf.drop_duplicates()
df_gnaf

X_test = df_gnaf[["floors_count", "unit_per_floor", "weighted_average_price", "area"]]
X_test

price_est2 = price_estimation(X_test)
price_est2

df_gnaf = df_gnaf.reset_index()
df_gnaf = pd.concat([df_gnaf, price_est2], axis=1)
df_gnaf


df_gnaf["estimated_price_m2"] = np.exp(df_gnaf["y_pred"])
df_gnaf["estimated_price"] = df_gnaf["estimated_price_m2"] * df_gnaf["area"]
df_gnaf

df_gnaf.to_csv(r"d:\Users\A.Fallah\Desktop\Price Estimation Model\Last Round\GNAF_Estimation_New.csv", index=False)
