==> 3.Customers in Tehran.ipynb <==
import pandas as pd

excel_file_path = r"d:\Users\A.Fallah\Desktop\Price Estimation Model\DimCustomer.csv"
df_dim = pd.read_csv(excel_file_path)
df_dim

excel_file_path = r"d:\Users\A.Fallah\Desktop\Price Estimation Model\VerifiedAddress1.csv"
df_verify = pd.read_csv(excel_file_path)
df_verify = df_verify.rename(columns={'Customerkey': 'customerkey'})
df_verify = df_verify.drop_duplicates()
df_verify

import numpy as np 

df_customer = pd.merge(df_dim, df_verify, on='customerkey', how='inner')
df_customer['homePostalCode'] = df_customer['homePostalCode'].astype(np.int64)
df_customer = df_customer[df_customer['homePostalCode'] > 1111111111]
df_customer

df_customer.to_csv(r"d:\Users\A.Fallah\Desktop\Price Estimation Model\Last Round\Customers_in_Tehran_New.csv", index=False)
