==> 4.Customers' House Estimation GNAF.ipynb <==
### Loading Data

import pandas as pd

excel_file_path = r"d:\Users\A.Fallah\Desktop\Price Estimation Model\Last Round\GNAF_Estimation_New.csv"
df_allestimation = pd.read_csv(excel_file_path)
df_allestimation['postal_code'] = df_allestimation['postal_code'].astype(str)
df_allestimation['Tours'] = df_allestimation['postal_code'].apply(lambda x: x[:5])
df_allestimation

a = sum(df_allestimation['estimated_price']) / sum(df_allestimation['area'])
a

### Summary Stats on All Estimation

df = df_allestimation.copy()
grouped_df = df.groupby(['neighborhood', 'District'])
df['record_count'] = grouped_df['estimated_price_m2'].transform('count')
df['update_fee'] = df['area'] * df['estimated_price_m2']
df['weighted_average_price_m2'] = grouped_df['update_fee'].transform('sum') / (grouped_df['area'].transform('sum') * 1000000)
df['average_house_price'] = grouped_df['update_fee'].transform('sum') / (grouped_df['estimated_price_m2'].transform('count') * 1000000000)
select = ['neighborhood', 'District', 'weighted_average_price_m2', 'average_house_price', 'record_count']
df_stat = df[select]
df_stat = df_stat.drop_duplicates()
df_stat = df_stat.sort_values('weighted_average_price_m2', ascending=False)
df_stat = df_stat.reset_index()
df_stat

df_stat.to_csv(r"d:\Users\A.Fallah\Desktop\Price Estimation Model\Last Round\Neighborhood_Stats.csv",  index=False)

### Finding Non Business 

import pandas as pd

excel_file_path = r"d:\Users\A.Fallah\Desktop\Price Estimation Model\NonBussiness\GNAF.csv"
df_gnaf = pd.read_csv(excel_file_path)
df_gnaf = df_gnaf.rename(columns={'postalcode': 'postal_code'})
df_gnaf

import pandas as pd

excel_file_path = r"d:\Users\A.Fallah\Desktop\Price Estimation Model\Last Round\Customers_in_Tehran_New.csv"
df_customers = pd.read_csv(excel_file_path)
df_customers = df_customers.rename(columns={'homePostalCode': 'postal_code'})
df_customers

df_customers_gnaf = pd.merge(df_customers, df_gnaf, on='postal_code', how='inner')
df_customers_gnaf

sel = ['customerkey', 'postal_code']
df_customers_gnaf = df_customers_gnaf[sel]
df_customers_gnaf

df_cust_select = df_customers_gnaf[['postal_code']].drop_duplicates()
df_cust_select['postal_code'] = df_cust_select['postal_code'].astype(str)
df_all = pd.merge(df_cust_select, df_allestimation, on='postal_code', how='inner')
df_all = df_all.drop_duplicates()
df_all

df_all.to_csv(r"d:\Users\A.Fallah\Desktop\Price Estimation Model\Last Round\Tehran_Estimation_house.csv",  index=False)

df_2 = df_all.copy()
grouped_df = df_2.groupby(['neighborhood', 'District'])
df_2['record_count'] = grouped_df['estimated_price_m2'].transform('count')
df_2['update_fee'] = df_2['area'] * df_2['estimated_price_m2']
df_2['weighted_average_price_m2'] = grouped_df['update_fee'].transform('sum') / (grouped_df['area'].transform('sum') * 1000000)
df_2['average_house_price'] = grouped_df['update_fee'].transform('sum') / (grouped_df['estimated_price_m2'].transform('count') * 1000000000)
select = ['neighborhood', 'District', 'weighted_average_price_m2', 'average_house_price']
df_stat_2 = df_2[select]
df_stat_2 = df_stat_2.drop_duplicates()
df_stat_2 = df_stat_2.sort_values('weighted_average_price_m2', ascending=False)
df_stat_2 = df_stat_2.reset_index()
df_stat_2

df_stat_2.to_csv(r"d:\Users\A.Fallah\Desktop\Price Estimation Model\Last Round\Neighborhood_Stats_Mofid.csv",  index=False)
