==> 6.Correlation-House & Portfo.ipynb <==
import pandas as pd

excel_file_path = r"d:\Users\A.Fallah\Desktop\Price Estimation Model\Last Round\Tehran_Estimation_Customers.csv"
df = pd.read_csv(excel_file_path)
df = df.drop_duplicates()
df

import pandas as pd

excel_file_path = r"d:\Users\A.Fallah\Desktop\Price Estimation Model\Estimation Outputs\Customers_in_Tehran_New.csv"
df_verify = pd.read_csv(excel_file_path)
sel = ['customerkey', 'homePostalCode']
df_verify = df_verify[sel]
df_verify

excel_file_path = r"d:\Users\A.Fallah\Desktop\Price Estimation Model\AveragePortfo-Verified.csv"
df_portfo = pd.read_csv(excel_file_path)
df_portfo = df_portfo.rename(columns={'customerKey': 'customerkey'})
df_portfo

df_mofid = pd.merge(df_verify, df_portfo, on='customerkey', how='outer')
df_mofid = df_mofid.dropna(subset=['homePostalCode'])
df_mofid = df_mofid.fillna(0)
df_mofid

import pandas as pd

excel_file_path = r"d:\Users\A.Fallah\Desktop\Fund 2\Final Table\MF for House Model.csv"
df_MF = pd.read_csv(excel_file_path)
df_MF
 

df_MF = df_MF.set_index('customerKey')
df_MF['average_mf'] = df_MF.mean(axis=1)



df_MF = df_MF[['average_mf']]
df_MF = df_MF.reset_index()
df_MF = df_MF.rename(columns={'customerKey': 'customerkey'})
df_MF

df_mutual = pd.merge(df_verify, df_MF, on='customerkey', how='inner')
sel = ['customerkey', 'average_mf']
df_mutual = df_mutual[sel]
df_mutual

df_final_portfo = pd.merge(df_mofid, df_mutual, on='customerkey', how='outer')
df_final_portfo = df_final_portfo.fillna(0)
df_final_portfo['average_portfo'] = df_final_portfo['average'].add(df_final_portfo['average_mf'])
df_final_portfo

merged_df = pd.merge(df, df_final_portfo, on='customerkey', how='inner')
merged_df 

df_x = merged_df.copy()
df_x = df_x[(df_x['estimated_price']>15*1000*1000*1000) & (df_x['average_portfo']>50*1000*1000)]
df_x

30000 * 500 * 1000 * 1000 * 0.04 

correlation = merged_df['estimated_price'].corr(merged_df['average_portfo'])
correlation

import numpy as np 

merged_df['averageT'] = merged_df['average_portfo'] / 10

percentile_1 = np.percentile(merged_df['averageT'], 1)
percentile_25 = np.percentile(merged_df['averageT'], 25)
percentile_50 = np.percentile(merged_df['averageT'], 50)
percentile_75 = np.percentile(merged_df['averageT'], 75)
percentile_90 = np.percentile(merged_df['averageT'], 90)
percentile_95 = np.percentile(merged_df['averageT'], 95)
percentile_99 = np.percentile(merged_df['averageT'], 99)

description_with_percentiles = merged_df['averageT'].describe(percentiles=[0.01, 0.25, 0.5, 0.75, 0.9, 0.95, 0.99])
description_with_percentiles.loc['1%',] = percentile_1
description_with_percentiles.loc['25%',] = percentile_25
description_with_percentiles.loc['50%',] = percentile_50
description_with_percentiles.loc['75%',] = percentile_75
description_with_percentiles.loc['90%',] = percentile_90
description_with_percentiles.loc['95%',] = percentile_95
description_with_percentiles.loc['99%',] = percentile_99
description_with_percentiles

import numpy as np  

percentile_1 = np.percentile(merged_df['estimated_price'], 1)
percentile_25 = np.percentile(merged_df['estimated_price'], 25)
percentile_50 = np.percentile(merged_df['estimated_price'], 50)
percentile_75 = np.percentile(merged_df['estimated_price'], 75)
percentile_90 = np.percentile(merged_df['estimated_price'], 90)
percentile_95 = np.percentile(merged_df['estimated_price'], 95)
percentile_99 = np.percentile(merged_df['estimated_price'], 99)

description_with_percentiles = merged_df['estimated_price'].describe(percentiles=[0.01, 0.25, 0.5, 0.75, 0.9, 0.95, 0.99])
description_with_percentiles.loc['1%',] = percentile_1
description_with_percentiles.loc['25%',] = percentile_25
description_with_percentiles.loc['50%',] = percentile_50
description_with_percentiles.loc['75%',] = percentile_75
description_with_percentiles.loc['90%',] = percentile_90
description_with_percentiles.loc['95%',] = percentile_95
description_with_percentiles.loc['99%',] = percentile_99
description_with_percentiles

import pandas as pd

average_portfo_values = [500000000, 
                         1000000000, 
                         2500000000, 
                         5000000000]
prediction_thresholds = [5000000000, 
                         10000000000, 
                         15000000000, 
                         30000000000]


# Initialize an empty dictionary to store results
result_matrix = {}

# Iterate through each combination
for prediction in prediction_thresholds:
    for portfo_value in average_portfo_values:

        merged_df['portfo'] = merged_df['average'].apply(lambda x: 'low' if x < portfo_value else 'High')    
        merged_df['house'] = merged_df['estimated_price'].apply(lambda x: 'low' if x < prediction else 'High')
        fisher_table = merged_df[['house','portfo']]

        contingency_table = pd.crosstab(fisher_table['portfo'], fisher_table['house'])

        # Calculate total number of observations
        total_obs = contingency_table.values.sum()

        # Calculate probabilities for each cell
        probabilities = contingency_table / total_obs

        x = probabilities.iloc[0,0]/(probabilities.iloc[0,0]+probabilities.iloc[1,0])
        x = round(x,3)*100

        # Store the result in the dictionary
        result_matrix[(prediction, portfo_value)] = x

# Create a DataFrame from the dictionary
result_df = pd.DataFrame(result_matrix, index=['x']).T
result_df.index.names = ['Prediction Threshold', 'Average Portfolio']

# Print the result
result_df

result_df.to_csv(r"d:\Users\A.Fallah\Desktop\Price Estimation Model\Last Round\fisher2.csv", index=True)
